Move SQSWorker to common as Trait
